more fixes for uint32
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@13358 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/sound/loader/mod/moddata.cpp
@@ -334,7 +334,7 @@ public:
     SCF_CONSTRUCT_EMBEDDED_IBASE (scfiComponent);
   }
 
-  virtual csPtr<iSoundData> LoadSound (void *Buffer, unsigned long Size)
+  virtual csPtr<iSoundData> LoadSound (void *Buffer, uint32 Size)
   {
     csModSoundData *sd=NULL;
     if (csModSoundData::IsMod (Buffer, Size))
